git-svn-id: https://svn.tdr.iem.uni-due.de:8443/svn/sctp/trunk/rsplib2@808 516812f1-d3fd-0310-9ea4-9aededebcb87

+++ rsplib/registrar.c
@@ -3137,9 +3137,9 @@ static size_t registrarGetReportFunction(
    poolElements = ST_CLASS(poolHandlespaceManagementGetPoolElements)(&registrar->Handlespace);
    snprintf(statusText, CSPR_STATUS_SIZE,
             "%u PEs in %u Pool%s, %u Peer%s",
-            poolElements,
-            pools, (pools == 1) ? "" : "s",
-            peers, (peers == 1) ? "" : "s");
+            (unsigned int)poolElements,
+            (unsigned int)pools, (pools == 1) ? "" : "s",
+            (unsigned int)peers, (peers == 1) ? "" : "s");
    getComponentLocation(componentLocation, registrar->ASAPSocket, 0);
 
    *workload    = -1.0;
